Properly convert the markdown to HTML.
In 1d2448593b01e1b85bf426c577f92cf59b53f836 I fixed Shellac so
that the files are included in the artifact we distribute to NPM.
In so doing I discovered that the mechanism I was using to strip
the markdown wasn't working. I think I "tested" by retrieving it
from the CDN, which returned an old version (as the prior bug
explained).

This time I took a look at what's being published to verify the
changes. This should do the trick and makes the regexp a bit
simpler by doing it in two passes.

After this goes out the door I'll publish another version to
fully verify that shellac is working, end to end.